# src/main/java/com/vass/shop/infrastructure/adapters/inbound/rest/api/CartsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.23.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.vass.shop.infrastructure.adapters.inbound.rest.api;

import com.vass.shop.infrastructure.adapters.inbound.rest.dto.CartCreationRequest;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.CartItemRequest;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.CartResponse;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.CheckoutResponse;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ErrorModel;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-08T22:59:31.432094100+03:00[Europe/Bucharest]", comments = "Generator version: 7.23.0")
@Validated
@Tag(name = "carts", description = "the carts API")
public interface CartsApi {

    String PATH_ADD_ITEM_TO_CART = "/carts/{cartId}/items";
    /**
     * POST /carts/{cartId}/items : Add an item to the cart
     *
     * @param cartId  (required)
     * @param cartItemRequest  (required)
     * @return Item added or updated (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "addItemToCart",
        summary = "Add an item to the cart",
        responses = {
            @ApiResponse(responseCode = "200", description = "Item added or updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CartsApi.PATH_ADD_ITEM_TO_CART,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<CartResponse> addItemToCart(
        @Parameter(name = "cartId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cartId") UUID cartId,
        @Parameter(name = "CartItemRequest", description = "", required = true) @Valid @RequestBody CartItemRequest cartItemRequest
    );


    String PATH_CHECKOUT_CART = "/carts/{cartId}/checkout";
    /**
     * POST /carts/{cartId}/checkout : Finalize cart and persist to historical purchase data
     *
     * @param cartId  (required)
     * @return Checkout finalized (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "checkoutCart",
        summary = "Finalize cart and persist to historical purchase data",
        responses = {
            @ApiResponse(responseCode = "200", description = "Checkout finalized", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CheckoutResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CartsApi.PATH_CHECKOUT_CART,
        produces = { "application/json" }
    )
    ResponseEntity<CheckoutResponse> checkoutCart(
        @Parameter(name = "cartId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cartId") UUID cartId
    );


    String PATH_CREATE_CART = "/carts";
    /**
     * POST /carts : Initialize a new shopping cart
     *
     * @param cartCreationRequest  (required)
     * @return Cart created (status code 201)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "createCart",
        summary = "Initialize a new shopping cart",
        responses = {
            @ApiResponse(responseCode = "201", description = "Cart created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CartsApi.PATH_CREATE_CART,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<CartResponse> createCart(
        @Parameter(name = "CartCreationRequest", description = "", required = true) @Valid @RequestBody CartCreationRequest cartCreationRequest
    );


    String PATH_DELETE_CART = "/carts/{cartId}";
    /**
     * DELETE /carts/{cartId} : Clear or delete a cart
     *
     * @param cartId  (required)
     * @return Cart deleted successfully (status code 204)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "deleteCart",
        summary = "Clear or delete a cart",
        responses = {
            @ApiResponse(responseCode = "204", description = "Cart deleted successfully"),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CartsApi.PATH_DELETE_CART,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteCart(
        @Parameter(name = "cartId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cartId") UUID cartId
    );


    String PATH_GET_CART = "/carts/{cartId}";
    /**
     * GET /carts/{cartId} : Get cart state with computed pricing rules
     *
     * @param cartId  (required)
     * @return Detailed cart details (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "getCart",
        summary = "Get cart state with computed pricing rules",
        responses = {
            @ApiResponse(responseCode = "200", description = "Detailed cart details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CartsApi.PATH_GET_CART,
        produces = { "application/json" }
    )
    ResponseEntity<CartResponse> getCart(
        @Parameter(name = "cartId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cartId") UUID cartId
    );


    String PATH_REMOVE_ITEM_FROM_CART = "/carts/{cartId}/items/{productId}";
    /**
     * DELETE /carts/{cartId}/items/{productId} : Remove a product completely or decrease quantity from cart
     *
     * @param cartId  (required)
     * @param productId  (required)
     * @return Item removed successfully (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "removeItemFromCart",
        summary = "Remove a product completely or decrease quantity from cart",
        responses = {
            @ApiResponse(responseCode = "200", description = "Item removed successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CartsApi.PATH_REMOVE_ITEM_FROM_CART,
        produces = { "application/json" }
    )
    ResponseEntity<CartResponse> removeItemFromCart(
        @Parameter(name = "cartId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cartId") UUID cartId,
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    );

}